goldilocks: Goldilocks Adaptive Trial Designs for Time-to-Event Endpoints

Implements the Goldilocks adaptive trial design for a time to event outcome using a piecewise exponential model and conjugate Gamma prior distributions. The method closely follows the article by Broglio and colleagues doi:10.1080/10543406.2014.888569 , which allows users to explore the operating characteristics of different trial designs.

See also

Author

Maintainer: Graeme L. Hickey graemeleehickey@gmail.com (ORCID)

Authors:

Ying Wan ying.wan@bd.com

Thevaa Chandereng tc3123@cumc.columbia.edu (ORCID) (bayesDP code as a template)

Other contributors:

Becton, Dickinson and Company [copyright holder]

Tim Kacprowski t.kacprowski@tu-braunschweig.de (For code from fastlogrank R package.) [contributor]